+++ docs/release_notes.rst
@@ -11,6 +11,10 @@ Major Updates
 
 * Support for Python 2.6 has been dropped (#206, #217).
 
+Bug Fixes
+
+* Made parser more robust to bad source files (#168, #214)
+
 1.1.1 - October 4th, 2016
 -------------------------